Send time and timezone as part of register payload
